Fix warnings

+++ src/org/wfrobotics/reuse/hardware/sensors/MagnetoPotSRX.java
@@ -4,12 +4,12 @@ import com.ctre.CANTalon;
 import com.ctre.CANTalon.FeedbackDevice;
 /**
  * Class to use a Magnetic Potentiometer through a Talon SRX
- * like an analog potentiometer, when it doesn't have the 
+ * like an analog potentiometer, when it doesn't have the
  * full range of 0 to 1.
  * Created for using the 6127V1A360L.5FS (987-1393-ND on DigiKey)
  */
-public class MagnetoPotSRX extends MagnetoPot {
-    
+public class MagnetoPotSRX extends MagnetoPot
+{
     public CANTalon m_Talon;
 
     /**
@@ -30,7 +30,7 @@ public class MagnetoPotSRX extends MagnetoPot {
     {
         this(talon, fullRange, 0);
     }
-    
+
     /**
      * Initialize a new Magnetic Potentiometer through the SRX data port
      * @param talon TalonSRX object
@@ -40,19 +40,16 @@ public class MagnetoPotSRX extends MagnetoPot {
     public MagnetoPotSRX(CANTalon talon, double fullRange, double offset)
     {
         super(fullRange, offset);
-        
+
         m_Talon = talon;
         m_Talon.setFeedbackDevice(FeedbackDevice.AnalogPot);
     }
 
     public double getRawInput()
     {
-        return (double)m_Talon.getAnalogInRaw()/1023.0;
+        return m_Talon.getAnalogInRaw() / 1023.0;
     }
 
     @Override
-    public void free()
-    {
-        
-    }
+    public void free() { }
 }
