+++ src/org/wfrobotics/reuse/utilities/FastTrig.java
@@ -6,22 +6,22 @@ package org.wfrobotics.reuse.utilities;
 public abstract class FastTrig
 {
     private static final double[] sine = new double[360];
-    
+
     static
     {
         double sum = 0;
-        
+
         for (int index = 0; index < sine.length; index++)
         {
             sine[index] = Math.sin((double) index / sine.length * Math.PI * 2);
         }
-        
+
         // Exact Cartesian coordinates
         sine[0] = 0;
-        sine[(int) (sine.length * 1 / 4)] = 1;
-        sine[(int) (sine.length * 2 / 4)] = 0;
-        sine[(int) (sine.length * 3 / 4)] = -1;
-        
+        sine[sine.length * 1 / 4] = 1;
+        sine[sine.length * 2 / 4] = 0;
+        sine[sine.length * 3 / 4] = -1;
+
         // Condition the cache - Actually this improves performance much more than the lookup table
         for (int index = 0; index < sine.length; index++)
         {
@@ -29,20 +29,20 @@ public abstract class FastTrig
         }
         System.out.println("Fast trig cache conditioned: " + (sum < .000001));  // Don't optimize me out Java!
     }
-    
+
     public static double sin(double angle)
     {
-        int nearestAngle = (int) ((angle > 0) ? (angle + 0.5d) : (angle - 0.5d));        
+        int nearestAngle = (int) ((angle > 0) ? (angle + 0.5d) : (angle - 0.5d));
         int wrappedToIndexRange = ((nearestAngle % 360) + 360) % 360;
-        
+
         return sine[wrappedToIndexRange];
     }
-    
+
     public static double cos(double angle)
     {
         return sin(angle + 90);
     }
-    
+
     public static void debug()
     {
         long start;
@@ -51,23 +51,23 @@ public abstract class FastTrig
         long durCos;
         long durSin2;
         double sum = 0;
-        
+
         start = System.nanoTime();
         for (double index = -362.49; index < 362.5; index += .5)
         {
             sum += Math.sin(index);
-//            System.out.format("%.2f: %.9f%n", index, Math.sin(index));
+            //            System.out.format("%.2f: %.9f%n", index, Math.sin(index));
         }
         System.out.println(sum);
         sum = 0;
         durMathSin = System.nanoTime() - start;
-        
+
         System.out.println("Fast Sin:");
         start = System.nanoTime();
         for (double index = -362.49; index < 362.5; index += .5)
         {
             sum += sin(index);
-//            System.out.format("%.2f: %.9f%n", index, sin(index));
+            //            System.out.format("%.2f: %.9f%n", index, sin(index));
         }
         System.out.println(sum);
         sum = 0;
@@ -78,23 +78,23 @@ public abstract class FastTrig
         for (double index = -2.49; index < 362.5; index += .5)
         {
             sum += sin(index);
-//            System.out.format("%.2f: %.9f%n", index, cos(index));
+            //            System.out.format("%.2f: %.9f%n", index, cos(index));
         }
         System.out.println(sum);
         sum = 0;
         durCos = System.nanoTime() - start;
-        
+
         System.out.println("Fast Sin 2:");
         start = System.nanoTime();
         for (double index = -362.49; index < 362.5; index += .5)
         {
             sum += sin(index);
-//            System.out.format("%.2f: %.9f%n", index, sin(index));
+            //            System.out.format("%.2f: %.9f%n", index, sin(index));
         }
         System.out.println(sum);
         sum = 0;
         durSin2 = System.nanoTime() - start;
-        
+
         System.out.println("Math Sin Duration: " + durMathSin / 1000 + " us");
         System.out.println("Fast Sin Duration: " + durSin / 1000 + " us");
         System.out.println("Fast Cos Duration: " + durCos / 1000 + " us");
